# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.2.1"
  description: "API made for the SHCITY project"
  title: SHCity API
  contact: 
    name: "Alexandre Dias"
    url: "https://www.linkedin.com/in/alexandredias95/"
    email: "alex.m.a.dias95@gmail.com" 
# during dev, should point to your local machine
host: 193.136.127.202:80
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /data_type/add: 
    x-swagger-router-controller: data_type
    post:
      summary: Add a data_type.
      tags:
              - data_type
      consumes:
        - application/json
      operationId: data_type_add
      parameters:
        - in: body
          name: Data_type
          description: The Data_type to add.
          schema:
            $ref: "#/definitions/Data_type"
          required: true
      responses:
        201:
          description: Created
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        503:
          description: Error
          schema:
            $ref: "#/definitions/HelloWorldResponse"  
...
  /gates/gate_add: 
    x-swagger-router-controller: gates
    post:
      summary: Add a Gate.
      tags:
              - gates
      consumes:
        - application/json
      operationId: gate_add
      parameters:
        - in: body
          name: Gate
          description: The Gate to create.
          schema:
            $ref: "#/definitions/Gate"
          required: true
      responses:
        201:
          description: Created
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        503:
          description: Error
          schema:
            $ref: "#/definitions/HelloWorldResponse"
  /gates/gate_vehicle_affluence_add: 
    x-swagger-router-controller: gates
    post:
      summary: Add a Gate value.
      tags:
              - gates
      consumes:
        - application/json
      operationId: gate_vehicle_affluence_add
      parameters:
        - in: body
          name: Gate
          description: The Gate to add value.
          schema:
            $ref: "#/definitions/Gate_affluence"
          required: true
      responses:
        201:
          description: Created
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        503:
          description: Error
          schema:
            $ref: "#/definitions/HelloWorldResponse"
...
definitions:
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  Data_type:
    properties:  
        username:
          type : string
        token:
          type : string
        data_type_name:
          type : string
        description:
          type : string  
        unit:
          type : string
...
  Gate:
    properties:  
        username:
          type : string
        token:
          type : string
        gate_name:
          type : string
        description:
          type : string  
  Gate_affluence:
    properties:  
        username:
          type : string
        token:
          type : string
        gate_name:
          type : string
        schedule:
          type : string
        flow_cars:
          type : string
        date:
          type : string
...